# laravel/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Hospital Management API",
        "description": "API Documentation for the Hospital Management System",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Local API Server"
        }
    ],
    "paths": {
...
        "/patients": {
            "get": {
                "tags": [
                    "Patients"
                ],
                "summary": "Get list of patients",
                "description": "Returns list of patients",
                "operationId": "getPatientsList",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "post": {
                "tags": [
                    "Patients"
                ],
                "summary": "Create a new patient",
                "description": "Creates a new patient record",
                "operationId": "storePatient",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nom",
                                    "prenom",
                                    "date_naissance",
                                    "sexe",
                                    "adresse",
                                    "telephone"
                                ],
                                "properties": {
                                    "nom": {
                                        "type": "string",
                                        "example": "Doe"
                                    },
                                    "prenom": {
                                        "type": "string",
                                        "example": "John"
                                    },
                                    "date_naissance": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "1990-01-01"
                                    },
                                    "sexe": {
                                        "type": "string",
                                        "example": "M"
                                    },
                                    "adresse": {
                                        "type": "string",
                                        "example": "123 Main St"
                                    },
                                    "telephone": {
                                        "type": "string",
                                        "example": "0600000000"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "john@example.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Patient created"
                    }
                }
            }
        },
        "/prescriptions": {
            "get": {
                "tags": [
                    "Prescriptions"
                ],
                "summary": "Get list of prescriptions",
                "description": "Returns list of prescriptions",
                "operationId": "getPrescriptionsList",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "post": {
                "tags": [
                    "Prescriptions"
                ],
                "summary": "Create a new prescription",
                "description": "Creates a new prescription record",
                "operationId": "storePrescription",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "patient_id",
                                    "medecin_id",
                                    "date_prescription",
                                    "medications",
                                    "dosage",
                                    "frequency",
                                    "duration"
                                ],
                                "properties": {
                                    "patient_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "medecin_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "date_prescription": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-05-01"
                                    },
                                    "medications": {
                                        "type": "string",
                                        "example": "Paracetamol"
                                    },
                                    "dosage": {
                                        "type": "string",
                                        "example": "1000mg"
                                    },
                                    "frequency": {
                                        "type": "string",
                                        "example": "3 times a day"
                                    },
                                    "duration": {
                                        "type": "string",
                                        "example": "5 days"
                                    },
                                    "notes": {
                                        "type": "string",
                                        "example": "Take after meals"
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "active"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Prescription created"
                    }
                }
            }
        },
...
        "/users": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get list of users",
                "description": "Returns list of users",
                "operationId": "getUsersList",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Create a new user",
                "description": "Creates a new user record",
                "operationId": "storeUser",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Admin User"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "admin@hospital.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "secret123"
                                    },
                                    "role": {
                                        "type": "string",
                                        "example": "admin"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User created"
                    }
                }
            }
        }
    },
...
}